Updated with week 10 + 11
And also some constant declaration back from week 2.

I think week 10 is supposed to work, but I find it a little hard to debug.
I will have to implement the tree for week 11. It shouldn't be that bad, because I don't need to give it any functions or any real "abilties".